# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BorderCityBoxing API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
...
  /attendance/checkin:
    post:
      operationId: checkInToClass
      tags: [attendance]
      summary: Check in to a booked class
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInRequest"
      responses:
        "201":
          description: Checked in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResponse"
        "400":
          description: No active booking for this class
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: Already checked in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /attendance/my:
    get:
      operationId: getMyCheckIns
      tags: [attendance]
      summary: Get all classes the current user has checked into
      parameters:
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        "200":
          description: List of check-ins
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyCheckInsResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: Opaque session token — `Bearer <sid>`.
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: [string, 'null']
          format: email
        firstName:
          type: [string, 'null']
        lastName:
          type: [string, 'null']
        profileImageUrl:
          type: [string, 'null']
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: 'null'
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    CheckInRequest:
      type: object
      required: [classId]
      properties:
        classId:
          type: integer
    CheckInResponse:
      type: object
      required: [id, userId, classId, checkedInAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        classId:
          type: integer
        checkedInAt:
          type: string
          format: date-time
    MyCheckInsResponseItem:
      type: object
      required: [classId, checkedInAt]
      properties:
        classId:
          type: integer
        checkedInAt:
          type: string
          format: date-time
    MyCheckInsResponse:
      type: array
      items:
        $ref: "#/components/schemas/MyCheckInsResponseItem"